BruTile and BruTile.MbTiles to netstandard 2.0 and delete Desktop and Desktop.Cache projects
