+++ .github/workflows/build.yml
@@ -1,4 +1,4 @@
-name: ASP.NET Core CI
+name: Build BruTile
 
 on: [push]
