+++ BruTile-Build.sln
@@ -9,10 +9,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BruTile", "BruTile\BruTile.
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BruTile.MbTiles", "BruTile.MbTiles\BruTile.MbTiles.csproj", "{CFD89526-A4FB-4C52-A8D6-7F59A0E3EF7F}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BruTile.Desktop", "BruTile.Desktop\BruTile.Desktop.csproj", "{1C06F6B2-0218-4FC6-AF26-FE9076C6563C}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BruTile.Desktop.DbCache", "BruTile.Desktop.DbCache\BruTile.Desktop.DbCache.csproj", "{A2C5E5C9-D88E-4108-BD2F-08D6B4755967}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Performance", "Performance", "{FC9F55A8-498E-4FFE-9127-E8CB7725DC9D}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tools", "Tools", "{7DF3BF5B-CC8A-42E1-A616-F96924AB04BB}"
@@ -24,8 +20,6 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BruTile.Tests", "Tests\BruTile.Tests\BruTile.Tests.csproj", "{65B03D41-85CD-492C-99F9-4CD545651A3A}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BruTile.MbTiles.Tests", "Tests\BruTile.MbTiles.Tests\BruTile.MbTiles.Tests.csproj", "{1A6D2993-C380-4DF4-9952-A9F22FFD34EB}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BruTile.Desktop.Tests", "Tests\BruTile.Desktop.Tests\BruTile.Desktop.Tests.csproj", "{A3881F28-9986-4FD1-BB79-74171C3A1ADF}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -61,30 +55,6 @@ Global
 		{CFD89526-A4FB-4C52-A8D6-7F59A0E3EF7F}.Release|x64.Build.0 = Release|Any CPU
 		{CFD89526-A4FB-4C52-A8D6-7F59A0E3EF7F}.Release|x86.ActiveCfg = Release|Any CPU
 		{CFD89526-A4FB-4C52-A8D6-7F59A0E3EF7F}.Release|x86.Build.0 = Release|Any CPU
-		{1C06F6B2-0218-4FC6-AF26-FE9076C6563C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1C06F6B2-0218-4FC6-AF26-FE9076C6563C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1C06F6B2-0218-4FC6-AF26-FE9076C6563C}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{1C06F6B2-0218-4FC6-AF26-FE9076C6563C}.Debug|x64.Build.0 = Debug|Any CPU
-		{1C06F6B2-0218-4FC6-AF26-FE9076C6563C}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{1C06F6B2-0218-4FC6-AF26-FE9076C6563C}.Debug|x86.Build.0 = Debug|Any CPU
-		{1C06F6B2-0218-4FC6-AF26-FE9076C6563C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1C06F6B2-0218-4FC6-AF26-FE9076C6563C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1C06F6B2-0218-4FC6-AF26-FE9076C6563C}.Release|x64.ActiveCfg = Release|Any CPU
-		{1C06F6B2-0218-4FC6-AF26-FE9076C6563C}.Release|x64.Build.0 = Release|Any CPU
-		{1C06F6B2-0218-4FC6-AF26-FE9076C6563C}.Release|x86.ActiveCfg = Release|Any CPU
-		{1C06F6B2-0218-4FC6-AF26-FE9076C6563C}.Release|x86.Build.0 = Release|Any CPU
-		{A2C5E5C9-D88E-4108-BD2F-08D6B4755967}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A2C5E5C9-D88E-4108-BD2F-08D6B4755967}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A2C5E5C9-D88E-4108-BD2F-08D6B4755967}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{A2C5E5C9-D88E-4108-BD2F-08D6B4755967}.Debug|x64.Build.0 = Debug|Any CPU
-		{A2C5E5C9-D88E-4108-BD2F-08D6B4755967}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{A2C5E5C9-D88E-4108-BD2F-08D6B4755967}.Debug|x86.Build.0 = Debug|Any CPU
-		{A2C5E5C9-D88E-4108-BD2F-08D6B4755967}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A2C5E5C9-D88E-4108-BD2F-08D6B4755967}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A2C5E5C9-D88E-4108-BD2F-08D6B4755967}.Release|x64.ActiveCfg = Release|Any CPU
-		{A2C5E5C9-D88E-4108-BD2F-08D6B4755967}.Release|x64.Build.0 = Release|Any CPU
-		{A2C5E5C9-D88E-4108-BD2F-08D6B4755967}.Release|x86.ActiveCfg = Release|Any CPU
-		{A2C5E5C9-D88E-4108-BD2F-08D6B4755967}.Release|x86.Build.0 = Release|Any CPU
 		{13A6F9A1-6FE9-4BC9-B017-DCA436D06485}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{13A6F9A1-6FE9-4BC9-B017-DCA436D06485}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{13A6F9A1-6FE9-4BC9-B017-DCA436D06485}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -133,18 +103,6 @@ Global
 		{1A6D2993-C380-4DF4-9952-A9F22FFD34EB}.Release|x64.Build.0 = Release|Any CPU
 		{1A6D2993-C380-4DF4-9952-A9F22FFD34EB}.Release|x86.ActiveCfg = Release|Any CPU
 		{1A6D2993-C380-4DF4-9952-A9F22FFD34EB}.Release|x86.Build.0 = Release|Any CPU
-		{A3881F28-9986-4FD1-BB79-74171C3A1ADF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A3881F28-9986-4FD1-BB79-74171C3A1ADF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A3881F28-9986-4FD1-BB79-74171C3A1ADF}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{A3881F28-9986-4FD1-BB79-74171C3A1ADF}.Debug|x64.Build.0 = Debug|Any CPU
-		{A3881F28-9986-4FD1-BB79-74171C3A1ADF}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{A3881F28-9986-4FD1-BB79-74171C3A1ADF}.Debug|x86.Build.0 = Debug|Any CPU
-		{A3881F28-9986-4FD1-BB79-74171C3A1ADF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A3881F28-9986-4FD1-BB79-74171C3A1ADF}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A3881F28-9986-4FD1-BB79-74171C3A1ADF}.Release|x64.ActiveCfg = Release|Any CPU
-		{A3881F28-9986-4FD1-BB79-74171C3A1ADF}.Release|x64.Build.0 = Release|Any CPU
-		{A3881F28-9986-4FD1-BB79-74171C3A1ADF}.Release|x86.ActiveCfg = Release|Any CPU
-		{A3881F28-9986-4FD1-BB79-74171C3A1ADF}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -154,7 +112,6 @@ Global
 		{F1C36376-58EE-457F-A2A2-ADE35A39FC8D} = {FC9F55A8-498E-4FFE-9127-E8CB7725DC9D}
 		{65B03D41-85CD-492C-99F9-4CD545651A3A} = {D70B7185-B90D-4CB7-8FDB-441FADD9821D}
 		{1A6D2993-C380-4DF4-9952-A9F22FFD34EB} = {D70B7185-B90D-4CB7-8FDB-441FADD9821D}
-		{A3881F28-9986-4FD1-BB79-74171C3A1ADF} = {D70B7185-B90D-4CB7-8FDB-441FADD9821D}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {B183B66E-60DD-4958-AC91-81285E29B11D}
