--- BruTile.Desktop.DbCache/BruTile.Desktop.DbCache.csproj
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <SolutionDir Condition=" '$(SolutionDir)' == '' ">$(ProjectDir)..\</SolutionDir>
-    <TargetFrameworks>netstandard2.0</TargetFrameworks>
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <Description>Extensions for BruTile that require the full .Net Framework</Description>
-    <PackageTags>tiling gis osm geo file ado.net</PackageTags>
-    <RootNamespace>BruTile.Cache</RootNamespace>
-  </PropertyGroup>
-
-  <Import Project="$(SolutionDir)BruTile.Common.props" />
-
-  <ItemGroup>
-    <ProjectReference Include="..\BruTile\BruTile.csproj" />
-  </ItemGroup>
-
-</Project>
