+++ BruTile.sln
@@ -12,8 +12,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "00_Common", "00_Common", "{
 		README.md = README.md
 	EndProjectSection
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BruTile.Desktop", "BruTile.Desktop\BruTile.Desktop.csproj", "{2F6D9E1C-2001-40C8-89BA-FDF8597BB719}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BruTile", "BruTile\BruTile.csproj", "{9FB60E2D-C682-4A01-9584-8B495B8DA85A}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "03_Samples", "03_Samples", "{136D7621-C2BA-419F-AF8E-07BF1D2BCE1F}"
@@ -28,8 +26,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "02_Tests", "02_Tests", "{3B
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BruTile.Tests", "Tests\BruTile.Tests\BruTile.Tests.csproj", "{DFFD61EA-2A22-48D7-A5F7-17554AB10BCE}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BruTile.Desktop.Tests", "Tests\BruTile.Desktop.Tests\BruTile.Desktop.Tests.csproj", "{C2F848EF-E0AE-4A0C-909B-C9C10AFB2CEC}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BruTile.Samples.MbTiles", "Samples\BruTile.Samples.MbTiles\BruTile.Samples.MbTiles.csproj", "{F450520B-00BE-4FA0-9B40-8AAD5A74F321}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BruTile.MbTiles.Tests", "Tests\BruTile.MbTiles.Tests\BruTile.MbTiles.Tests.csproj", "{2DC1700A-F83F-44D9-A2AC-EA2DCF4E747F}"
@@ -43,8 +39,6 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BruTile.MbTiles", "BruTile.MbTiles\BruTile.MbTiles.csproj", "{D67D3DCC-6494-4C0A-91DD-DB9AE9EA43F1}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "01_Source", "01_Source", "{19780978-0590-4DCF-85A6-5D2941A1085C}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BruTile.Desktop.DbCache", "BruTile.Desktop.DbCache\BruTile.Desktop.DbCache.csproj", "{EA5DFD5A-BBF6-404F-A95E-DF096B56C5C0}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "05_Tools", "05_Tools", "{6EE2299B-A487-4B50-BEE3-3CEE6D07A6C5}"
 EndProject
@@ -63,10 +57,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{2F6D9E1C-2001-40C8-89BA-FDF8597BB719}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2F6D9E1C-2001-40C8-89BA-FDF8597BB719}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2F6D9E1C-2001-40C8-89BA-FDF8597BB719}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2F6D9E1C-2001-40C8-89BA-FDF8597BB719}.Release|Any CPU.Build.0 = Release|Any CPU
 		{9FB60E2D-C682-4A01-9584-8B495B8DA85A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{9FB60E2D-C682-4A01-9584-8B495B8DA85A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9FB60E2D-C682-4A01-9584-8B495B8DA85A}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -87,10 +77,6 @@ Global
 		{DFFD61EA-2A22-48D7-A5F7-17554AB10BCE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DFFD61EA-2A22-48D7-A5F7-17554AB10BCE}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DFFD61EA-2A22-48D7-A5F7-17554AB10BCE}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C2F848EF-E0AE-4A0C-909B-C9C10AFB2CEC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C2F848EF-E0AE-4A0C-909B-C9C10AFB2CEC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C2F848EF-E0AE-4A0C-909B-C9C10AFB2CEC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C2F848EF-E0AE-4A0C-909B-C9C10AFB2CEC}.Release|Any CPU.Build.0 = Release|Any CPU
 		{F450520B-00BE-4FA0-9B40-8AAD5A74F321}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{F450520B-00BE-4FA0-9B40-8AAD5A74F321}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F450520B-00BE-4FA0-9B40-8AAD5A74F321}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -111,10 +97,6 @@ Global
 		{D67D3DCC-6494-4C0A-91DD-DB9AE9EA43F1}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D67D3DCC-6494-4C0A-91DD-DB9AE9EA43F1}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D67D3DCC-6494-4C0A-91DD-DB9AE9EA43F1}.Release|Any CPU.Build.0 = Release|Any CPU
-		{EA5DFD5A-BBF6-404F-A95E-DF096B56C5C0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EA5DFD5A-BBF6-404F-A95E-DF096B56C5C0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EA5DFD5A-BBF6-404F-A95E-DF096B56C5C0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EA5DFD5A-BBF6-404F-A95E-DF096B56C5C0}.Release|Any CPU.Build.0 = Release|Any CPU
 		{DB13792C-8391-4D3A-AC4A-DA910E9464DC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{DB13792C-8391-4D3A-AC4A-DA910E9464DC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DB13792C-8391-4D3A-AC4A-DA910E9464DC}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -128,19 +110,16 @@ Global
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{2F6D9E1C-2001-40C8-89BA-FDF8597BB719} = {19780978-0590-4DCF-85A6-5D2941A1085C}
 		{9FB60E2D-C682-4A01-9584-8B495B8DA85A} = {19780978-0590-4DCF-85A6-5D2941A1085C}
 		{57766550-A465-4797-80B7-9D7746D2DCF3} = {136D7621-C2BA-419F-AF8E-07BF1D2BCE1F}
 		{24A0D64A-F894-4EB4-8CB7-8F6081BFE324} = {136D7621-C2BA-419F-AF8E-07BF1D2BCE1F}
 		{11C2C7B4-3A50-4984-9466-478526091EE2} = {136D7621-C2BA-419F-AF8E-07BF1D2BCE1F}
 		{DFFD61EA-2A22-48D7-A5F7-17554AB10BCE} = {3B8883BB-764B-40AA-B7BD-A30F4C9D5808}
-		{C2F848EF-E0AE-4A0C-909B-C9C10AFB2CEC} = {3B8883BB-764B-40AA-B7BD-A30F4C9D5808}
 		{F450520B-00BE-4FA0-9B40-8AAD5A74F321} = {136D7621-C2BA-419F-AF8E-07BF1D2BCE1F}
 		{2DC1700A-F83F-44D9-A2AC-EA2DCF4E747F} = {3B8883BB-764B-40AA-B7BD-A30F4C9D5808}
 		{ADE23B05-8CCE-4EAD-BCA6-0A469F60B29B} = {136D7621-C2BA-419F-AF8E-07BF1D2BCE1F}
 		{7114313B-7563-4D86-B052-5E5667D7E17C} = {E8E1A22E-2F90-42E1-A8B6-BAB783881FD3}
 		{D67D3DCC-6494-4C0A-91DD-DB9AE9EA43F1} = {19780978-0590-4DCF-85A6-5D2941A1085C}
-		{EA5DFD5A-BBF6-404F-A95E-DF096B56C5C0} = {19780978-0590-4DCF-85A6-5D2941A1085C}
 		{DB13792C-8391-4D3A-AC4A-DA910E9464DC} = {6EE2299B-A487-4B50-BEE3-3CEE6D07A6C5}
 		{ACB6A591-8735-4BB1-B8DC-4FBA65C4D722} = {136D7621-C2BA-419F-AF8E-07BF1D2BCE1F}
 	EndGlobalSection
