--- BruTile.v2.ncrunchsolution
@@ -1,13 +1,0 @@
-<SolutionConfiguration>
-  <FileVersion>1</FileVersion>
-  <AutoEnableOnStartup>True</AutoEnableOnStartup>
-  <AllowParallelTestExecution>false</AllowParallelTestExecution>
-  <AllowTestsToRunInParallelWithThemselves>true</AllowTestsToRunInParallelWithThemselves>
-  <FrameworkUtilisationTypeForNUnit>UseDynamicAnalysis</FrameworkUtilisationTypeForNUnit>
-  <FrameworkUtilisationTypeForGallio>UseStaticAnalysis</FrameworkUtilisationTypeForGallio>
-  <FrameworkUtilisationTypeForMSpec>UseStaticAnalysis</FrameworkUtilisationTypeForMSpec>
-  <FrameworkUtilisationTypeForMSTest>UseStaticAnalysis</FrameworkUtilisationTypeForMSTest>
-  <EngineModes>Run all tests automatically:BFRydWU=;Run all tests manually:BUZhbHNl;Run impacted tests automatically, others manually (experimental!):CklzSW1wYWN0ZWQ=;Run pinned tests automatically, others manually:CElzUGlubmVk</EngineModes>
-  <MetricsExclusionList>
-</MetricsExclusionList>
-</SolutionConfiguration>
