--- BruTile.v3.ncrunchsolution
@@ -1,14 +1,0 @@
-﻿<SolutionConfiguration>
-  <Settings>
-    <AllowParallelTestExecution>False</AllowParallelTestExecution>
-    <SolutionConfigured>True</SolutionConfigured>
-  </Settings>
-  <EngineModes>
-    <EngineMode>
-      <Name>Run impacted tests automatically, others manually (experimental!)</Name>
-      <Settings>
-        <TestsToExecuteAutomatically>IsImpacted</TestsToExecuteAutomatically>
-      </Settings>
-    </EngineMode>
-  </EngineModes>
-</SolutionConfiguration>
