+++ Tests/BruTile.Tests/BruTile.Tests.csproj
@@ -16,6 +16,7 @@
   </ItemGroup>
   
   <ItemGroup>
+    <PackageReference Include="System.Data.OleDb" Version="5.0.0" />
     <PackageReference Include="System.Net.Requests" Version="4.3.0" />
     <PackageReference Include="System.Net.Http" Version="4.3.0" />
     <PackageReference Include="System.Threading.Tasks.Parallel" Version="4.3.0" />
